% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloom.R
\name{pev_data_bloom}
\alias{pev_data_bloom}
\alias{pev_data_bloom.default}
\alias{pev_data_bloom.character}
\alias{pev_data_bloom.pev_fcont}
\alias{pev_data_bloom.pev_fdisc}
\title{Get HCL data for a series of hex-colors}
\usage{
pev_data_bloom(hex, include_cvd = TRUE, ...)

\method{pev_data_bloom}{default}(hex, ...)

\method{pev_data_bloom}{character}(hex, n = NULL, include_cvd = TRUE,
  ...)

\method{pev_data_bloom}{pev_fcont}(hex, n = NULL, include_cvd = TRUE,
  ...)

\method{pev_data_bloom}{pev_fdisc}(hex, n = NULL, include_cvd = TRUE,
  ...)
}
\arguments{
\item{hex}{\code{character} vector of hex-colors}

\item{include_cvd}{\code{logical}, indicates to include data for
for color-vision deficiency}

\item{...}{other args (not used)}

\item{n}{\code{integer} number of colors to use for unbounded palette-functions}
}
\value{
\code{data.frame} with variables \code{cvd}, \code{x}, \code{hex}, \code{hue}, \code{chroma}, \code{luminance}
}
\description{
Get HCL data for a series of hex-colors
}
