% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{pev_data_distance}
\alias{pev_data_distance}
\title{Get perceptual-distance of continuous-palette from reference color}
\usage{
pev_data_distance(.fcont, hex_ref = NULL, n_panel = 40,
  include_cvd = TRUE)
}
\arguments{
\item{.fcont}{\code{object} that can be coerced to \code{pev_fcont},
when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{hex_ref}{\code{character} single hex-colors to act as the reference, default value
is the beginning of the range of the continuous-palette, i.e. at \code{x = 0}.}

\item{n_panel}{\code{numeric} number of panels into which to discretize the
domain (0 to 1) of the continous-palette.}

\item{include_cvd}{\code{logical}, indicates to include data for
for color-vision deficiency}
}
\value{
\code{data.frame} with variables \code{cvd}, \code{x}, \code{hex}, \code{hex_ref}, \code{distance}
}
\description{
Get perceptual-distance of continuous-palette from reference color
}
\examples{
  pev_data_distance("Purple-Green")
}
