% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl-data.R
\name{pev_data_hcl}
\alias{pev_data_hcl}
\alias{pev_data_hcl.default}
\alias{pev_data_hcl.character}
\alias{pev_data_hcl.pev_fcont}
\alias{pev_data_hcl.pev_fdisc}
\title{Get HCL data on series of hex-colors}
\usage{
pev_data_hcl(hex, ...)

\method{pev_data_hcl}{default}(hex, ...)

\method{pev_data_hcl}{character}(hex, n = NULL, include_cvd = TRUE,
  ...)

\method{pev_data_hcl}{pev_fcont}(hex, n = NULL, include_cvd = TRUE,
  ...)

\method{pev_data_hcl}{pev_fdisc}(hex, n = NULL, include_cvd = TRUE,
  ...)
}
\arguments{
\item{hex}{\code{character} vector of hex-colors}

\item{...}{other args (not used)}

\item{n}{\code{integer} number of colors to use for unbounded palette-functions}

\item{include_cvd}{\code{logical}, indicates to include data for
for color-vision deficiency}
}
\value{
\code{data.frame} with variables \code{cvd}, \code{x}, \code{hex},
\code{hue}, \code{chroma}, \code{luminance}, \code{is_rgb_limit}
}
\description{
TODO: some notes here on \code{is_rgb_limit} being applied to
non-cvd case.
}
\examples{
  pev_data_hcl("Viridis")
  pev_data_hcl(grDevices::heat.colors(10))
}
