% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev_data_hcl_ref.R
\name{pev_data_hcl_ref}
\alias{pev_data_hcl_ref}
\alias{pev_data_hcl_ref.default}
\alias{pev_data_hcl_ref.character}
\alias{pev_data_hcl_ref.pev_fcont}
\alias{pev_data_hcl_ref.pev_funbounded}
\alias{pev_data_hcl_ref.pev_fbounded}
\title{Get minimum perceptual-distance from reference color}
\usage{
pev_data_hcl_ref(.fpal, hex_ref, n = NULL, method = "cie2000",
  include_cvd = TRUE, ...)

\method{pev_data_hcl_ref}{default}(.fpal, hex_ref, n = NULL,
  method = "cie2000", include_cvd = TRUE, ...)

\method{pev_data_hcl_ref}{character}(.fpal, hex_ref, n = NULL,
  method = "cie2000", include_cvd = TRUE, ...)

\method{pev_data_hcl_ref}{pev_fcont}(.fpal, hex_ref, n = NULL,
  method = "cie2000", include_cvd = TRUE, ...)

\method{pev_data_hcl_ref}{pev_funbounded}(.fpal, hex_ref, n = NULL,
  method = "cie2000", include_cvd = TRUE, ...)

\method{pev_data_hcl_ref}{pev_fbounded}(.fpal, hex_ref, n = NULL,
  method = "cie2000", include_cvd = TRUE, ...)
}
\arguments{
\item{.fpal}{palette-function.}

\item{hex_ref}{\code{character} vector of hex-colors}

\item{n}{\code{integer}, number of points to discretize the range (0, 1)
of the palette-function to determine the minimum perceptual-distance.}

\item{method}{\code{character} method to use for distance calculation,
passed to \code{farver::compare_color()}.
One of: \code{"euclidean"}, \code{"cie1976"}, \code{"cie94"}, \code{"cie2000"}, or \code{"cmc"}.}

\item{include_cvd}{\code{logical}, indicates to include data for
for color-vision deficiency}

\item{...}{other args (not used)}
}
\value{
\code{data.frame} with variables \code{cvd}, \code{x_nearest}, \code{distance_nearest},
\code{hex}, \code{hex_ref}, \code{hue}, \code{chroma}, \code{luminance}.
}
\description{
An approximation, for now.
}
\examples{
  pev_data_hcl_ref("Viridis", "#008F97")
}
