% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont}
\alias{pev_fcont}
\alias{pev_fcont_div}
\alias{pev_fcont_rescale}
\alias{pev_fcont_reverse}
\alias{pev_fcont_cvd}
\title{Create continuous-palette function}
\usage{
pev_fcont(hcl_param)

pev_fcont_div(pev_fcont_low, pev_fcont_high)

pev_fcont_rescale(pev_fcont, limits = c(0, 1))

pev_fcont_reverse(pev_fcont)

pev_fcont_cvd(pev_fcont, type = c("deutan", "protan", "tritan", "none"),
  severity = 1)
}
\arguments{
\item{hcl_param}{Object with S3 class \code{pev_hcl_param},
created using \code{\link[=pev_hcl_param]{pev_hcl_param()}}.}

\item{pev_fcont_low}{\code{function}, similar to \code{pev_fcont}, used for the
low end of a diverging scale. Low end of this scale becomes the
middle of the diverging scale.}

\item{pev_fcont_high}{\code{function}, similar to \code{pev_fcont}, used for the
high end of a diverging scale. Low end of this scale becomes the
middle of the diverging scale.}

\item{pev_fcont}{\code{function}, when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{limits}{\code{numeric} vector of length 2, rescaling parameters. The input
values to the returned function, \code{c(0, 1)}, are mapped to \code{limits} as
inputs to the input function.}

\item{type}{\code{character}, describes color-vision deficiency. One of
\code{"deutan"}, \code{"protan"}, \code{"tritan"}, \code{"none"}.}

\item{severity}{\code{numeric}, number between 0 (none) and 1 to describe
the severity of color-vision deficiency.}
}
\value{
\code{function}, when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.
}
\description{
Create continuous-palette function
}
\examples{
  # Define HCL parameters for palette functions
  pals <- pev_map_hcl_param(
    colorspace::hcl_palettes(palette = c("Purples 3", "Greens 3"))
  )

  # Create palette functions
  fcont_purple <- pev_fcont(pals[["Purples 3"]])
  fcont_green <- pev_fcont(pals[["Greens 3"]])
  fcont_purple
  fcont_green

  # Modify palette functions
  pev_fcont_reverse(fcont_purple)
  pev_fcont_rescale(fcont_purple, limits = c(0.25, 0.75))

  # Create diverging-palette function
  fcont_purple_green <- pev_fcont_div(fcont_purple, fcont_green)
  fcont_purple_green

  # Modify palette function to simulate color-vision deficiency
  pev_fcont_cvd(fcont_purple_green, type = "deutan")
}
