% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_cvd}
\alias{pev_fcont_cvd}
\title{Create continuous-palette function from HCL parameters}
\usage{
pev_fcont_cvd(pev_fcont, type = c("deutan", "protan", "tritan", "none"),
  severity = 1)
}
\arguments{
\item{pev_fcont}{\code{function}, when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{type}{\code{character}, describes color-vision deficiency. One of
\code{"deutan"}, \code{"protan"}, \code{"tritan"}, \code{"none"}.}

\item{severity}{\code{numeric}, number between 0 (none) and 1 to describe
the severity of color-vision deficiency.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
Create continuous-palette function from HCL parameters
}
