% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_diverging}
\alias{pev_fcont_diverging}
\title{Create continuous-palette function from HCL parameters}
\usage{
pev_fcont_diverging(pev_fcont_low, pev_fcont_high)
}
\arguments{
\item{pev_fcont_low}{\code{function}, similar to \code{pev_fcont}, used for the
low end of a diverging scale. Low end of this scale becomes the
middle of the diverging scale.}

\item{pev_fcont_high}{\code{function}, similar to \code{pev_fcont}, used for the
high end of a diverging scale. Low end of this scale becomes the
middle of the diverging scale.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
Create continuous-palette function from HCL parameters
}
