% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_hcl}
\alias{pev_fcont_hcl}
\title{Create continuous-palette function from HCL parameters}
\usage{
pev_fcont_hcl(hcl_param)
}
\arguments{
\item{hcl_param}{Object with S3 class \code{pev_hcl},
created using \code{\link[=pev_hcl]{pev_hcl()}}.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
Create continuous-palette function from HCL parameters
}
