% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_rescale}
\alias{pev_fcont_rescale}
\title{Create continuous-palette function from HCL parameters}
\usage{
pev_fcont_rescale(pev_fcont, limits = c(0, 1))
}
\arguments{
\item{pev_fcont}{\code{function}, when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{limits}{\code{numeric} vector of length 2, rescaling parameters.
The input values to the returned function, \code{c(0, 1)},
are mapped to \code{limits} as inputs to the input function.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
Create continuous-palette function from HCL parameters
}
