% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{pev_fcont_reverse}
\alias{pev_fcont_reverse}
\alias{pev_fcont_reverse.default}
\alias{pev_fcont_reverse.pev_fcont}
\alias{pev_fdisc_reverse}
\alias{pev_fdisc_reverse.default}
\alias{pev_fdisc_reverse.pev_funbounded}
\alias{pev_fdisc_reverse.pev_fbounded}
\title{Reverse palette-function}
\usage{
pev_fcont_reverse(.fcont, ...)

\method{pev_fcont_reverse}{default}(.fcont, ...)

\method{pev_fcont_reverse}{pev_fcont}(.fcont, ...)

pev_fdisc_reverse(.fdisc, ...)

\method{pev_fdisc_reverse}{default}(.fdisc, ...)

\method{pev_fdisc_reverse}{pev_funbounded}(.fdisc, ...)

\method{pev_fdisc_reverse}{pev_fbounded}(.fdisc, ...)
}
\arguments{
\item{.fcont}{\code{object} that can be coerced to \code{pev_fcont},
when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{...}{Other args (not used)}

\item{.fdisc}{\code{object} that can be coerced to \code{pev_fdisc},
when called with an integer, returns the corresponding (hex-code) values.}
}
\value{
Object of of the same type as \code{.fpal};
a palette function, with S3 class \code{pev_fcont} or \code{pev_fdisc}.
}
\description{
Reverse palette-function
}
\examples{
  # Create continuous diverging-palette function
  fcont_purple_green <- pev_fcont("Purple-Green")
  fcont_purple_green

  # Create discrete diverging-palette function
  fdisc_purple_green <- pev_fdisc(fcont_purple_green, n = 11)
  fdisc_purple_green

  # Reverse the palettes
  pev_fcont_reverse(fcont_purple_green)
  pev_fdisc_reverse(fdisc_purple_green)

}
