% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev_fdisc.R
\name{pev_fdisc}
\alias{pev_fdisc}
\alias{pev_fdisc.default}
\alias{pev_fdisc.character}
\alias{pev_fdisc.pev_fcont}
\alias{pev_fdisc.pev_fdisc}
\title{Coerce to discrete-palette function}
\usage{
pev_fdisc(.fdisc, ...)

\method{pev_fdisc}{default}(.fdisc, ...)

\method{pev_fdisc}{character}(.fdisc, ...)

\method{pev_fdisc}{pev_fcont}(.fdisc, method = c("post", "panel"), ...)

\method{pev_fdisc}{pev_fdisc}(.fdisc, ...)
}
\arguments{
\item{.fdisc}{\code{object} that can be coerced to \code{pev_fdisc},
when called with an integer, returns the corresponding (hex-code) values.}

\item{...}{other arguments (not used).}

\item{method}{\code{character}, describes how the domain of the continuous palette
is to be discretized, can be \code{"post"} or \code{"panel"} (using a fencing analogy).}
}
\value{
\code{function} with S3 class \code{pev_disc},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
A discrete-palette function takes an integers, from 1 to
the maximum size of the palette, and returns a vector of character hex-codes.
You con use a discrete-palette function to
build a custom ggplot2 scale, using \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.
}
\details{
These functions help you build, modify, and compose discrete-palette
functions.

A discrete-palette function can be \emph{bounded} or \emph{unbounded}. A bounded
function has a maxiumum number of possible colors, and is created from a
series of hex-colors, for example,
\href{https://www.tableau.com/about/blog/2016/7/colors-upgrade-tableau-10-56782}{Tableau 10}.
An unbounded function is created from a continuous-palette function, thus has no maximum
number of possible colors.

A discrete-palette function can be constructed using \code{\link[=pev_fdisc]{pev_fdisc()}}; it takes
an argument \code{.fdisc}, which can be one of:

\describe{
\item{\code{character}}{A series of hex-codes,
returns a \strong{bounded} discrete-palette function.}
\item{\code{pev_fcont}}{A continuous-palette function, to be discretized
according to \code{method} (\code{"post"} or \code{"panel"}).
This returns an \strong{unbounded} discrete-palette function.}
\item{\code{pev_fdisc}}{If you provide a \code{pev_fdisc}, this is a no-op.}
}

The print method for a \code{pev_disc} function generates a plot of the palette.

The other functions that return continuous-palette functions are:

\describe{
\item{\code{\link[=pev_fdisc_cvd]{pev_fdisc_cvd()}}}{Modify output to simulate color-vision deficiency.}
\item{\code{\link[=pev_fdisc_reverse]{pev_fdisc_reverse()}}}{Reverse palette-function.}
}
}
\examples{
 # Use a colorspace palette
 pev_fdisc(pev_fcont("Blues 2"), n = 11, method = "post")

}
