% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-vision.R
\name{pev_fpal_cvd}
\alias{pev_fpal_cvd}
\alias{pev_fpal_cvd.default}
\alias{pev_fpal_cvd.pev_fcont}
\alias{pev_fpal_cvd.pev_fdisc}
\title{Modify output to simulate color-vision deficiency}
\usage{
pev_fpal_cvd(.fpal, ...)

\method{pev_fpal_cvd}{default}(.fpal, ...)

\method{pev_fpal_cvd}{pev_fcont}(.fpal, type = c("deutan", "protan",
  "tritan", "none"), severity = 1, ...)

\method{pev_fpal_cvd}{pev_fcont}(.fpal, type = c("deutan", "protan",
  "tritan", "none"), severity = 1, ...)

\method{pev_fpal_cvd}{pev_fdisc}(.fpal, type = c("deutan", "protan",
  "tritan", "none"), severity = 1, ...)
}
\arguments{
\item{.fpal}{Object with S3 class \code{pev_fcont} or \code{pev_fdisc},
a continuous or discrete palette-function}

\item{...}{Other args (not used)}

\item{type}{\code{character}, describes color-vision deficiency. One of
\code{"deutan"}, \code{"protan"}, \code{"tritan"}, \code{"none"}.}

\item{severity}{\code{numeric}, number between 0 (none) and 1 to describe
the severity of color-vision deficiency.}
}
\value{
Object of of the same type as \code{.fpal};
a palette function, with S3 class \code{pev_fcont} or \code{pev_fdisc}.
}
\description{
TODO: URL to explain CVD
}
\examples{
  # Create continuous diverging-palette function
  fcont_purple_green <- pev_fcont("Purple-Green")
  fcont_purple_green

  # Create discrete diverging-palette function
  fdisc_purple_green <- pev_fdisc(fcont_purple_green, n = 11)
  fdisc_purple_green

  # Simulate color-vision deficiency
  pev_fpal_cvd(fcont_purple_green, type = "deutan")
  pev_fpal_cvd(fdisc_purple_green, type = "deutan")

}
