% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{pev_fpal_reverse}
\alias{pev_fpal_reverse}
\alias{pev_fpal_reverse.default}
\alias{pev_fpal_reverse.pev_fcont}
\alias{pev_fpal_reverse.pev_fdisc}
\title{Reverse palette-function}
\usage{
pev_fpal_reverse(.fpal, ...)

\method{pev_fpal_reverse}{default}(.fpal, ...)

\method{pev_fpal_reverse}{pev_fcont}(.fpal, ...)

\method{pev_fpal_reverse}{pev_fdisc}(.fpal, ...)
}
\arguments{
\item{.fpal}{Object with S3 class \code{pev_fcont} or \code{pev_fdisc},
a continuous or discrete palette-function}

\item{...}{Other args (not used)}
}
\value{
Object of of the same type as \code{.fpal};
a palette function, with S3 class \code{pev_fcont} or \code{pev_fdisc}.
}
\description{
Reverse palette-function
}
\examples{
  # Create continuous diverging-palette function
  fcont_purple_green <- pev_fcont("Purple-Green")
  fcont_purple_green

  # Create discrete diverging-palette function
  fdisc_purple_green <- pev_fdisc(fcont_purple_green, n = 11)
  fdisc_purple_green

  # Simulate color-vision deficiency
  pev_fpal_reverse(fcont_purple_green)
  pev_fpal_reverse(fdisc_purple_green)

}
