% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_hex.R
\name{pev_fpal_to_hex}
\alias{pev_fpal_to_hex}
\alias{pev_fpal_to_hex.default}
\alias{pev_fpal_to_hex.pev_fcont}
\alias{pev_fpal_to_hex.pev_fdisc}
\title{Convert to hex-colors}
\usage{
pev_fpal_to_hex(.fpal, ...)

\method{pev_fpal_to_hex}{default}(.fpal, ...)

\method{pev_fpal_to_hex}{pev_fcont}(.fpal, n = 11, method = c("post",
  "panel"), ...)

\method{pev_fpal_to_hex}{pev_fdisc}(.fpal, ...)
}
\arguments{
\item{.fpal}{Object with S3 class \code{pev_fcont} or \code{pev_fdisc},
a continuous or discrete palette-function}

\item{...}{Other args (not used)}

\item{n}{coercible to \code{integer}, number colors to put in the palette.}

\item{method}{\code{character}, describes how the domain of the continuous palette
is to be discretized, can be \code{"post"} or \code{"panel"} (using a fencing analogy).}
}
\value{
\code{character} vector of hex-colors
}
\description{
Convert to hex-colors
}
\examples{
  # Create continuous diverging-palette function
  fcont_purple_green <- pev_fcont("Purple-Green")
  fcont_purple_green

  # Create discrete diverging-palette function
  fdisc_purple_green <- pev_fdisc(fcont_purple_green, n = 11)
  fdisc_purple_green

  # Convert to hex-colors
  pev_fpal_to_hex(fcont_purple_green, n = 11)
  pev_fpal_to_hex(fdisc_purple_green)

}
