% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bloom-gg.R
\name{pev_gg_bloom}
\alias{pev_gg_bloom}
\alias{pev_gg_bloom_target}
\alias{pev_gg_bloom_lum}
\title{ggplot for HCL data on series of hex-colors}
\usage{
pev_gg_bloom(data_bloom, label = FALSE, max_chroma = NULL)

pev_gg_bloom_target(data_bloom, label = FALSE, max_chroma = 200)

pev_gg_bloom_lum(data_bloom, label = FALSE)
}
\arguments{
\item{data_bloom}{\code{data.frame} created using \code{pev_data_bloom}.}

\item{label}{\code{logical} indicates to use axes labels, titles, etc.}

\item{max_chroma}{\code{numeric} upper-limit for the chroma scale}
}
\description{
Create a plot like Tableau made \href{https://www.tableau.com/about/blog/2016/7/colors-upgrade-tableau-10-56782}{here}.
}
\examples{
  # without color-vision deficiency, with labels
  data_no_cvd <- pev_data_bloom("Viridis", include_cvd = FALSE)
  pev_gg_bloom(data_no_cvd, label = TRUE)

  # without color-vision deficiency, without labels
  pev_data_bloom("Viridis") \%>\% pev_gg_bloom()
}
