% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative.R
\name{pev_gg_derivative}
\alias{pev_gg_derivative}
\title{ggplot for perceptual-derivative of continuous-palette}
\usage{
pev_gg_derivative(data_drv)
}
\arguments{
\item{data_drv}{\code{data.frame}, created using \code{pev_data_separation}.}
}
\value{
\code{ggplot} object
}
\description{
ggplot for perceptual-derivative of continuous-palette
}
\examples{
  data_drv <- pev_data_derivative("Viridis")
  pev_gg_derivative(data_drv)
}
