% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{pev_gg_distance}
\alias{pev_gg_distance}
\title{ggplot for perceptual-derivative of continuous-palette}
\usage{
pev_gg_distance(data_dist)
}
\arguments{
\item{data_dist}{\code{data.frame}, created using \code{\link[=pev_data_distance]{pev_data_distance()}}.}
}
\value{
\code{ggplot} object
}
\description{
ggplot for perceptual-derivative of continuous-palette
}
\examples{
  data_dist <- pev_data_distance("Purple-Green")
  pev_gg_distance(data_dist)
}
