% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl-bloom.R
\name{pev_gg_hcl_bloom}
\alias{pev_gg_hcl_bloom}
\alias{pev_gg_hcl_bloom_target}
\alias{pev_gg_hcl_bloom_lum}
\title{ggplot for HCL data on series of hex-colors}
\usage{
pev_gg_hcl_bloom(data_hcl, data_hcl_ref = NULL, label = FALSE,
  max_chroma = NULL)

pev_gg_hcl_bloom_target(data_hcl, data_hcl_ref = NULL, label = FALSE,
  max_chroma = NULL)

pev_gg_hcl_bloom_lum(data_hcl, data_hcl_ref = NULL, label = FALSE)
}
\arguments{
\item{data_hcl}{\code{data.frame} created using \code{pev_data_hcl}.}

\item{data_hcl_ref}{\code{data.frame} created using \code{pev_data_hcl_ref}.}

\item{label}{\code{logical} indicates to use axes labels, titles, etc.}

\item{max_chroma}{\code{numeric} upper-limit for the chroma scale.}
}
\value{
Object with S3 class \code{gg} (ggplot).
}
\description{
Create a plot like Tableau made \href{https://www.tableau.com/about/blog/2016/7/colors-upgrade-tableau-10-56782}{here}.
}
\examples{
  # without color-vision deficiency, with labels
  data_no_cvd <- pev_data_hcl("Viridis", include_cvd = FALSE)
  pev_gg_hcl_bloom(data_no_cvd, label = TRUE)

  # without color-vision deficiency, without labels
  pev_data_hcl("Viridis") \%>\% pev_gg_hcl_bloom()
}
