% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev_gg_hcl_plane.R
\name{pev_gg_hcl_plane}
\alias{pev_gg_hcl_plane}
\title{ggplot for HCL data on series of hex-colors}
\usage{
pev_gg_hcl_plane(data_hcl, data_hcl_ref = NULL)
}
\arguments{
\item{data_hcl}{\code{data.frame} created using \code{pev_data_hcl}.}

\item{data_hcl_ref}{\code{data.frame} created using \code{pev_data_hcl_ref}.}
}
\value{
Object with S3 class \code{gg} (ggplot).
}
\description{
ggplot for HCL data on series of hex-colors
}
\examples{
  pev_gg_hcl_plane(pev_data_hcl("Greens"))
}
