% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{pev_gg_separation}
\alias{pev_gg_separation}
\title{ggplot for perceptual-differences within palette}
\usage{
pev_gg_separation(data_sep, ncol = 2, height_tick = 1)
}
\arguments{
\item{data_sep}{\code{data.frame}, created using \code{pev_data_separation}.}

\item{ncol}{\code{numeric}, number of columns in the facet}

\item{height_tick}{\code{numeric}, height (units of \code{difference}) of the
cross-wise ticks}
}
\value{
\code{ggplot} object
}
\description{
ggplot for perceptual-differences within palette
}
\examples{
  fcont <- pev_fcont("Dynamic")
  fdisc <- pev_fdisc(fcont, n = 7, method = "panel")
  data_sep <- pev_data_separation(fdisc)
  pev_gg_separation(data_sep)
}
