% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_param.R
\name{pev_hcl_param}
\alias{pev_hcl_param}
\title{Create HCL parameter-set}
\usage{
pev_hcl_param(h1, h2 = h1, c1, c2 = c1, l1, l2 = l1, p1 = 1,
  p2 = 1, cmax = c1, fixup = TRUE)
}
\arguments{
\item{h1}{hue value in the HCL color description, has to be in [0, 360].}

\item{h2}{hue value in the HCL color description, has to be in [0, 360].}

\item{c1}{chroma value in the HCL color description.}

\item{c2}{chroma value in the HCL color description.}

\item{l1}{luminance value in the HCL color description.}

\item{l2}{luminance value in the HCL color description.}

\item{p1}{control parameter determining how chroma and luminance should
be increased (1 = linear, 2 = quadratic, etc.).}

\item{p2}{control parameter determining how chroma and luminance should
be increased (1 = linear, 2 = quadratic, etc.).}

\item{cmax}{Maximum chroma value in the HCL color description.}

\item{fixup}{logical. Should the color be corrected to a valid RGB value?}
}
\value{
Object with S3 class \code{pev_hcl_param}
}
\description{
This follows the colorspace scheme.
}
\examples{
 pev_hcl_param(h1 = 0, h2 = 360, c1 = 60, l1 = 60)
}
