% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex-distance.R
\name{pev_hex_derivative}
\alias{pev_hex_derivative}
\title{Calculate perceptual-derivative for sequence of hex-colors}
\usage{
pev_hex_derivative(hex, method = "cie2000")
}
\arguments{
\item{hex}{\code{character} vector of hex-colors}

\item{method}{\code{character} method to use for distance calculation,
passed to \code{farver::compare_color()}.
One of: \code{"euclidean"}, \code{"cie1976"}, \code{"cie94"}, \code{"cie2000"}, or \code{"cmc"}.}
}
\value{
\code{numeric}
}
\description{
This assumes that \code{hex} reperesents colors on a continuous-scale
where the domain varies uniformly from 0 to 1.
}
\examples{
  pev_fcont("Viridis")(seq(0, 1, by = 0.025)) \%>\%
    pev_hex_derivative()
}
