% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{pev_data_separation}
\alias{pev_data_separation}
\title{Get perceptual-separation within discrete-palette}
\usage{
pev_data_separation(.fdisc, n = NULL, method = "cie2000",
  include_cvd = TRUE)
}
\arguments{
\item{.fdisc}{\code{object} that can be coerced to \code{pev_fdisc},
when called with an integer, returns the corresponding (hex-code) values.}

\item{n}{\code{integer} number of colors to use for unbounded palette-functions}

\item{method}{\code{character}, describes how the domain of the continuous palette
is to be discretized, can be \code{"post"} or \code{"panel"} (using a fencing analogy).}

\item{include_cvd}{\code{logical} or \code{character}, indicates to include data for
for color-vision deficiency. Possible character values: \code{"none"}, \code{"protan"},
\code{"deutan"}, \code{"tritan"}.}
}
\value{
\code{tbl_df} with variables \code{cvd}, \code{hex_a}, \code{hex_b}, \code{distance}
}
\description{
Get perceptual-separation within discrete-palette
}
\examples{
  pev_fcont("Dynamic") \%>\%
    pev_fdisc(n = 7, method = "panel") \%>\%
    pev_data_separation()
}
\seealso{
pev_gg_separation
}
