% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{pev_data_separation}
\alias{pev_data_separation}
\title{Get separation within discrete-palette}
\usage{
pev_data_separation(.fdisc, n = NULL, method = "cie2000",
  include_cvd = TRUE)
}
\arguments{
\item{.fdisc}{\code{object} that can be coerced to \code{pev_fdisc},
when called with an integer, returns the corresponding (hex-code) values.}

\item{n}{\code{integer} number of colors to use for unbounded palette-functions}

\item{method}{\code{character}, describes how the domain of the continuous palette
is to be discretized, can be \code{"post"} or \code{"panel"} (using a fencing analogy).}

\item{include_cvd}{\code{logical}, indicates to include data for
for color-vision deficiency}
}
\value{
\code{tbl_df} with variables \code{cvd}, \code{hex_a}, \code{hex_b}, \code{distance}
}
\description{
Get separation within discrete-palette
}
\examples{
  fcont <- pev_fcont("Dynamic")
  fdisc <- pev_fdisc(fcont, n = 7, method = "panel")
  pev_data_separation(fdisc)
}
\seealso{
pev_gg_separation
}
