% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{pev_data_separation}
\alias{pev_data_separation}
\title{Get perceptual-differences within palette}
\usage{
pev_data_separation(pal)
}
\arguments{
\item{pal}{\code{character} (coerceable to vector of hex-codes), describes a
discrete color palette. Interpolation within a palette is not required to
be a well-formed idea.}
}
\value{
\code{tbl_df} with variables \code{color_a}, \code{color_b}, \code{difference}
}
\description{
Get perceptual-differences within palette
}
