% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev_fcont.R
\name{pev_fcont}
\alias{pev_fcont}
\alias{pev_fcont.default}
\alias{pev_fcont.pev_fcont}
\alias{pev_fcont.character}
\alias{pev_fcont.pev_hcl}
\title{Coerce to continuous-palette function}
\usage{
pev_fcont(.fcont, ...)

\method{pev_fcont}{default}(.fcont, ...)

\method{pev_fcont}{pev_fcont}(.fcont, ...)

\method{pev_fcont}{character}(.fcont, ...)

\method{pev_fcont}{pev_hcl}(.fcont, ...)
}
\arguments{
\item{.fcont}{\code{object} that can be coerced to \code{pev_fcont},
when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{...}{other arguments (not used).}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
A continuous-palette function takes a vector of numbers, each between
0 and 1, and returns a vector of character strings, each containing the
correpsonding hex-code. You con use a continuous-palette function to
build a custom ggplot2 scale, using \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}}.
}
\details{
These functions help you build, modify, and compose continuous-palette
functions. By working with functions, rather than with a finite set of
hex-colors, we can avoid interpolation-errors as we compose and rescale.

A continuous-palette function can be constructed using \code{\link[=pev_fcont]{pev_fcont()}}; it takes
a single argument \code{.fcont}, which can be one of:

\describe{
\item{\code{character}}{Name of a \code{palette} used by \link[colorspace:hcl_palettes]{colorspace::hcl_palettes}.}
\item{\code{pev_hcl}}{Set of HCL parameters returned by \code{\link[=pev_hcl]{pev_hcl()}}.}
\item{\code{pev_fcont}}{If you provide a \code{pev_fcont}, this is a no-op.}
}

The print method for a \code{pev_fcont} function generates a plot of the palette.

The other functions that return continuous-palette functions are:

\describe{
\item{\code{\link[=pev_fcont_cvd]{pev_fcont_cvd()}}}{Modify output to simulate color-vision deficiency.}
\item{\code{\link[=pev_fcont_diverging]{pev_fcont_diverging()}}}{Create a diverging-palette function from two functions.}
\item{\code{\link[=pev_fcont_rescale]{pev_fcont_rescale()}}}{Rescale input to continuous-palette function.}
\item{\code{\link[=pev_fcont_reverse]{pev_fcont_reverse()}}}{Reverse palette-function.}
}
}
\examples{
# Use a colorspace palette
pev_fcont("Blues 2")

}
