% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-vision.R
\name{pev_fcont_cvd}
\alias{pev_fcont_cvd}
\alias{pev_fcont_cvd.default}
\alias{pev_fcont_cvd.pev_fcont}
\alias{pev_fdisc_cvd}
\alias{pev_fdisc_cvd.default}
\alias{pev_fdisc_cvd.pev_funbounded}
\alias{pev_fdisc_cvd.pev_fbounded}
\title{Modify output to simulate color-vision deficiency}
\usage{
pev_fcont_cvd(.fcont, ...)

\method{pev_fcont_cvd}{default}(.fcont, ...)

\method{pev_fcont_cvd}{pev_fcont}(.fcont, type = c("deutan", "protan",
  "tritan", "none"), severity = 1, ...)

pev_fdisc_cvd(.fdisc, ...)

\method{pev_fdisc_cvd}{default}(.fdisc, ...)

\method{pev_fdisc_cvd}{pev_funbounded}(.fdisc, type = c("deutan",
  "protan", "tritan", "none"), severity = 1, ...)

\method{pev_fdisc_cvd}{pev_fbounded}(.fdisc, type = c("deutan", "protan",
  "tritan", "none"), severity = 1, ...)
}
\arguments{
\item{.fcont}{\code{object} that can be coerced to \code{pev_fcont},
when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{...}{Other args (not used)}

\item{type}{\code{character}, describes color-vision deficiency. One of
\code{"deutan"}, \code{"protan"}, \code{"tritan"}, \code{"none"}.}

\item{severity}{\code{numeric}, number between 0 (none) and 1 to describe
the severity of color-vision deficiency.}

\item{.fdisc}{\code{object} that can be coerced to \code{pev_fdisc},
when called with an integer, returns the corresponding (hex-code) values.}
}
\value{
Object of of the same type as \code{.fpal};
a palette function, with S3 class \code{pev_fcont} or \code{pev_fdisc}.
}
\description{
TODO: URL to explain CVD
}
\examples{
  # Create continuous diverging-palette function
  fcont_purple_green <- pev_fcont("Purple-Green")
  fcont_purple_green

  # Create discrete diverging-palette function
  fdisc_purple_green <- pev_fdisc(fcont_purple_green, n = 11)
  fdisc_purple_green

  # Simulate color-vision deficiency
  pev_fcont_cvd(fcont_purple_green, type = "deutan")
  pev_fdisc_cvd(fdisc_purple_green, type = "deutan")

}
