% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_cvd}
\alias{pev_fcont_cvd}
\title{Modify output to simulate color-vision deficiency}
\usage{
pev_fcont_cvd(.fcont, type = c("deutan", "protan", "tritan", "none"),
  severity = 1)
}
\arguments{
\item{.fcont}{\code{object} that can be coerced to \code{pev_fcont},
when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{type}{\code{character}, describes color-vision deficiency. One of
\code{"deutan"}, \code{"protan"}, \code{"tritan"}, \code{"none"}.}

\item{severity}{\code{numeric}, number between 0 (none) and 1 to describe
the severity of color-vision deficiency.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
TODO: URL to explain CVD
}
\examples{
  # Create diverging-palette function
  fcont_purple_green <- pev_fcont("Purple-Green")
  fcont_purple_green

  # Simulate color-vision deficiency
  pev_fcont_cvd(fcont_purple_green, type = "deutan")

}
