% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_diverging}
\alias{pev_fcont_diverging}
\title{Create continuous-palette function from HCL parameters}
\usage{
pev_fcont_diverging(.fcont_low, .fcont_high)
}
\arguments{
\item{.fcont_low}{\code{object} that can be coerced to \code{pev_fcont},
similar to \code{.fcont}, used for the low end of a diverging scale.
Low end of this scale becomes the middle of the diverging scale.}

\item{.fcont_high}{\code{object} that can be coerced to \code{pev_fcont},
similar to \code{.fcont}, used for the high end of a diverging scale.
Low end of this scale becomes the middle of the diverging scale.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
Create continuous-palette function from HCL parameters
}
\examples{
  # Create sequential-palette functions
  fcont_purple <- pev_fcont("Purples 3")
  fcont_green <- pev_fcont("Greens 3")

  # Create diverging-palette function
  pev_fcont_diverging(fcont_purple, fcont_green)
}
