% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_rescale}
\alias{pev_fcont_rescale}
\title{Rescale input to continuous-palette function}
\usage{
pev_fcont_rescale(.fcont, limits = c(0, 1))
}
\arguments{
\item{.fcont}{\code{object} that can be coerced to \code{pev_fcont},
when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}

\item{limits}{\code{numeric} vector of length 2, rescaling parameters.
The input values to the returned function, \code{c(0, 1)},
are mapped to \code{limits} as inputs to the input function.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
Rescale input to continuous-palette function
}
\examples{
  # Create sequential-palette function
  fcont_purple <- pev_fcont("Purples 3")
  fcont_purple

  # Rescale palette-function
  pev_fcont_rescale(fcont_purple, limits = c(0.25, 0.75))

}
