% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_reverse}
\alias{pev_fcont_reverse}
\title{Reverse palette-function}
\usage{
pev_fcont_reverse(.fcont)
}
\arguments{
\item{.fcont}{\code{object} that can be coerced to \code{pev_fcont},
when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
Reverse palette-function
}
\examples{
  # Create sequential-palette function
  fcont_purple <- pev_fcont("Purples 3")

  # Reverse palette-function
  pev_fcont_reverse(fcont_purple)

}
