% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{pev_fcont_reverse}
\alias{pev_fcont_reverse}
\title{Create continuous-palette function from HCL parameters}
\usage{
pev_fcont_reverse(pev_fcont)
}
\arguments{
\item{pev_fcont}{\code{function}, when called with a numeric vector with values
between 0 and 1, returns the corresponding (hex-code) values.}
}
\value{
\code{function} with S3 class \code{pev_fcont},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
Create continuous-palette function from HCL parameters
}
