% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pev_fdisc.R
\name{pev_fdisc}
\alias{pev_fdisc}
\alias{pev_fdisc.default}
\alias{pev_fdisc.character}
\alias{pev_fdisc.pev_fcont}
\alias{pev_fdisc.pev_fdisc}
\title{Coerce to discrete-palette function}
\usage{
pev_fdisc(.fdisc, ...)

\method{pev_fdisc}{default}(.fdisc, ...)

\method{pev_fdisc}{character}(.fdisc, ...)

\method{pev_fdisc}{pev_fcont}(.fdisc, n = 11, method = c("post",
  "panel"), ...)

\method{pev_fdisc}{pev_fdisc}(.fdisc, ...)
}
\arguments{
\item{.fdisc}{\code{object} that can be coerced to \code{pev_fdisc},
when called with an integer, returns the corresponding (hex-code) values.}

\item{...}{other arguments (not used).}

\item{n}{coercible to \code{integer}, number colors to put in the palette.}

\item{method}{\code{character}, describes how the domain of the continuous palette
is to be discretized, can be \code{"post"} or \code{"panel"} (using a fencing analogy).}
}
\value{
\code{function} with S3 class \code{pev_disc},
when called with a numeric vector with values between 0 and 1,
returns the corresponding (hex-code) values.
}
\description{
A discrete-palette function takes a vector of integers, from 1 to
the size of the palette, and returns a vector of character strings,
each containing the correpsonding hex-code.
You con use a discrete-palette function to
build a custom ggplot2 scale, using \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.
}
\details{
These functions help you build, modify, and compose discrete-palette
functions.

A discrete-palette function can be constructed using \code{\link[=pev_fdisc]{pev_fdisc()}}; it takes
an argument \code{.fdisc}, which can be one of:

\describe{
\item{\code{character}}{A series of hex-codes.}
\item{\code{pev_fcont}}{This will discretize a continuous palette into \code{n} colors,
according to the discretization \code{method} (\code{"post"} or \code{"panel"}).}
\item{\code{pev_fdisc}}{If you provide a \code{pev_fdisc}, this is a no-op.}
}

The print method for a \code{pev_disc} function generates a plot of the palette.

The other functions that return continuous-palette functions are:

\describe{
\item{\code{pev_fpal_cvd()}}{Modify output to simulate color-vision deficiency.}
\item{\code{pev_fpal_reverse()}}{Reverse palette-function.}
}
}
\examples{
 # Use a colorspace palette
 pev_fdisc(pev_fcont("Blues 2"), n = 11, method = "post")

}
