% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{pev_gg_separation}
\alias{pev_gg_separation}
\title{ggplot for erceptual-differences within palette}
\usage{
pev_gg_separation(data_sep)
}
\arguments{
\item{data_sep}{\code{data.frame} created using \code{pev_data_separation}}
}
\value{
\code{ggplot} object
}
\description{
ggplot for erceptual-differences within palette
}
\examples{
  library("colorspace")
  pal <- qualitative_hcl(n = 5)
  data_sep <- pev_data_separation(pal)
  pev_gg_separation(data_sep)
}
