% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\name{pev_gg_separation}
\alias{pev_gg_separation}
\title{ggplot for perceptual-distance within discrete-palette}
\usage{
pev_gg_separation(data_sep, ncol = 2, height_tick = 1)
}
\arguments{
\item{data_sep}{\code{data.frame}, created using \code{\link[=pev_data_separation]{pev_data_separation()}}.}

\item{ncol}{\code{numeric}, number of columns in the facet}

\item{height_tick}{\code{numeric}, height (units of \code{distance}) of the
cross-wise ticks}
}
\value{
\code{ggplot} object
}
\description{
ggplot for perceptual-distance within discrete-palette
}
\examples{
  data_sep <-
    pev_fcont("Dynamic") \%>\%
    pev_fdisc(n = 7, method = "panel") \%>\%
    pev_data_separation()
  pev_gg_separation(data_sep)
}
