% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcl_param.R
\name{pev_map_hcl_param}
\alias{pev_map_hcl_param}
\title{Create list of HCL parameter-sets}
\usage{
pev_map_hcl_param(hcl_palettes)
}
\arguments{
\item{hcl_palettes}{Object with S3 class \code{hcl_palettes},
created using \code{\link[colorspace:hcl_palettes]{colorspace::hcl_palettes()}}.}
}
\value{
Named list, elements are objects with S3 class \code{pev_hcl_param}
}
\description{
For each row in the \code{hcl_palettes} data frame
}
